+++ tests/Unit/Factory/ContentUpdateFactoryTest.php
@@ -37,7 +37,19 @@ class ContentUpdateFactoryTest extends TestCase
 
     private function getFactory(ObjectsInterface $objects): ContentUpdateFactory
     {
-        $configProvider = new ConfigProvider('host', 'key', 'key', 1, 1);
+        $configProvider = new ConfigProvider(
+            'config_name',
+            [
+                'config_name' => [
+                    'host' => 'host',
+                    'publicKey' => '',
+                    'privateKey' => '',
+                    'connectionTimeout' => 5.0,
+                    'requestTimeout' => 5.0,
+                ],
+            ]
+        );
+
         $serializer = $this->createMock(LuigisBoxSerializer::class);
         $serializer->expects($this->once())
             ->method('serialize')