+++ CHANGELOG.md
@@ -1,5 +1,29 @@
+v2.x
+===================
+* 2.0.0
+  * [BC BREAK] change default connection timeout from 10s to 4s, request timeout from 10s to 8s
+  * [BC BREAK] change bundle configuration:
+
+    from
+    ```php
+    answear_luigis_box:
+      publicKey: 'public'
+      privateKey: 'private'
+    ```
+    to
+    ```php
+    answear_luigis_box:
+      configs:
+          your_config_name:
+              publicKey: 'public'
+              privateKey: 'private'
+    ```
+    See [README](README.md) for more details.
+
 v1.x
 ===================
+* 1.4.0
+  * [BC BREAK] remove casting bool values in url filters.
 
 * 1.3.0
   * (feature) Update by query